chore: move discord channel funcs to utils
