+++ src/events/commands/channels/unlockChannel.ts
@@ -1,7 +1,7 @@
 import { ChannelType, PermissionFlagsBits } from 'discord.js';
 import { InteractionError } from '@structures/interactions/_Interaction';
 import { SlashCommand } from '@structures/interactions/SlashCommand';
-import { editChannelPermission } from '@controllers/channelController';
+import { editChannelPermission } from '@utils/discordChannels';
 
 export default new SlashCommand('unlock')
 	.setDescription('Allow a role to speak in a channel again')
